JHP/requirements.txt
git+https://github.com/openai/CLIP.git
faiss-cpu==1.8.0
Flask==2.2.2
Flask_Cors==4.0.1
numpy==1.24.3
openai==1.41.0
Pillow==10.4.0
torch==2.4.0
python-dotenv
MarkupSafe==2.1.1
Werkzeug==2.2.3
torchvision==0.19.0

